Add a button to recenter the view on the entire boundary
