+++ assets/README.md
@@ -1,0 +1,1 @@
+Material Icons courtesy https://fonts.google.com/icons
